requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
python-multipart==0.0.20
openpyxl==3.1.5

# Testing
pytest==8.3.3
pytest-mock==3.14.0
pytest-cov==5.0.0
pytest-asyncio==0.23.7
